Padronização do tipo das variáveis de id.

+++ calculadora/src/main/java/br/com/desafioCompasso/calculadora/controller/dto/GrupoMedicamentoDto.java
@@ -6,7 +6,7 @@ import br.com.desafioCompasso.calculadora.model.GrupoMedicamento;
 
 public class GrupoMedicamentoDto {
 	
-	private BigInteger id;
+	private Long id;
 	private String nome;
 	
 	
@@ -21,14 +21,16 @@ public class GrupoMedicamentoDto {
 		this.nome= grupoMed.getNome();
 	}
 	
-	public BigInteger getId() {
+
+	
+	public Long getId() {
 		return id;
 	}
-	
-	public void setId(BigInteger id) {
+
+	public void setId(Long id) {
 		this.id = id;
 	}
-	
+
 	public String getNome() {
 		return nome;
 	}
